# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerSummary } from "./playerSummary";

export interface AnalyticsSummary {
  totalDays: number;
  retailer: PlayerSummary;
  wholesaler: PlayerSummary;
  factory: PlayerSummary;
}