# lib/api-zod/src/generated/types/arrivalResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { GameEvent } from "./gameEvent";

export interface ArrivalResult {
  arrivedQuantity: number;
  newStock: number;
  holdingCostCharged: number;
  events: GameEvent[];
}