# lib/api-zod/src/generated/types/gameEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { GameEventData } from "./gameEventData";
import type { GameEventType } from "./gameEventType";
import type { PlayerRole } from "./playerRole";

export interface GameEvent {
  id: number;
  role: PlayerRole;
  type: GameEventType;
  message: string;
  day: number;
  acknowledged: boolean;
  data?: GameEventData;
}