# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStateTurnPhase } from "./gameStateTurnPhase";
import type { PlayerRole } from "./playerRole";
import type { PlayerState } from "./playerState";

export interface GameState {
  currentDay: number;
  currentTurnRole: PlayerRole;
  turnPhase: GameStateTurnPhase;
  players: PlayerState[];
  isGameOver: boolean;
  /** @nullable */
  winnerRole?: string | null;
}