# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";
import type { PlayerRole } from "./playerRole";

export interface Order {
  id: number;
  fromRole: PlayerRole;
  toRole: PlayerRole;
  quantity: number;
  status: OrderStatus;
  createdAtDay: number;
  estimatedArrivalDay: number;
  /** @nullable */
  deliveredAtDay?: number | null;
}