# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerRole } from "./playerRole";

export interface OrderInput {
  fromRole: PlayerRole;
  /** @minimum 0 */
  quantity: number;
}