# lib/api-zod/src/generated/types/playerRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */

export type PlayerRole = (typeof PlayerRole)[keyof typeof PlayerRole];

export const PlayerRole = {
  retailer: "retailer",
  wholesaler: "wholesaler",
  factory: "factory",
} as const;