# lib/api-zod/src/generated/types/playerState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerRole } from "./playerRole";

export interface PlayerState {
  role: PlayerRole;
  stock: number;
  money: number;
  backlog: number;
  pendingOrders: number;
  totalHoldingCost: number;
  totalStockoutCost: number;
}