# lib/api-zod/src/generated/types/playerSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerSummary {
  stock: number;
  money: number;
  backlog: number;
  totalHoldingCost: number;
  totalStockoutCost: number;
}