# lib/api-zod/src/generated/types/transitItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerRole } from "./playerRole";

export interface TransitItem {
  orderId: number;
  fromRole: PlayerRole;
  toRole: PlayerRole;
  quantity: number;
  createdAtDay: number;
  estimatedArrivalDay: number;
  turnsRemaining: number;
}