# lib/api-zod/src/generated/types/turnSnapshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Beer Game Supply Chain Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface TurnSnapshot {
  day: number;
  retailerStock: number;
  wholesalerStock: number;
  factoryStock: number;
  retailerMoney: number;
  wholesalerMoney: number;
  factoryMoney: number;
  retailerBacklog: number;
  wholesalerBacklog: number;
  factoryBacklog: number;
  demand: number;
}